refactor: Move csp_rdp from transport folder to src
csp if udp was nothing, and there never really came any other transport layer protocols. So put RDP back down to source level to simplify tree structure.
