+++ meson.build
@@ -31,9 +31,8 @@ csp_sources = files([
 	'src/crypto/csp_hmac.c',
 	'src/crypto/csp_sha1.c',
 	'src/crypto/csp_xtea.c',
-	'src/transport/csp_rdp.c',
-	'src/transport/csp_rdp_queue.c',
-	'src/transport/csp_udp.c',
+	'src/csp_rdp.c',
+	'src/csp_rdp_queue.c',
 	'src/csp_buffer.c',
 	'src/csp_bridge.c',
 	'src/csp_conn.c',
