+++ src/csp_rdp.c
@@ -7,7 +7,6 @@
  */
 
 #include "csp_rdp_queue.h"
-#include "csp_transport.h"
 
 #include <stdlib.h>
 #include <string.h>
@@ -20,9 +19,9 @@
 #include <csp/arch/csp_semaphore.h>
 #include <csp/arch/csp_time.h>
 
-#include "../csp_port.h"
-#include "../csp_conn.h"
-#include "../csp_io.h"
+#include "csp_port.h"
+#include "csp_conn.h"
+#include "csp_io.h"
 
 #define RDP_SYN 0x08
 #define RDP_ACK 0x04
@@ -1023,15 +1022,5 @@ void csp_rdp_get_opt(unsigned int * window_size, unsigned int * conn_timeout_ms,
 		*ack_delay_count = csp_rdp_ack_delay_count;
 }
 
-#if (CSP_HAVE_STDIO)
-void csp_rdp_conn_print(csp_conn_t * conn) {
-
-	if (conn == NULL)
-		return;
-
-	csp_print("\tRDP: S:%d (closed by 0x%x), rcv %u, snd %u, win %" PRIu32 "\n",
-		   conn->rdp.state, conn->rdp.closed_by, conn->rdp.rcv_cur, conn->rdp.snd_una, conn->rdp.window_size);
-}
-#endif
 
 #endif  // CSP_USE_RDP
