--- src/transport/CMakeLists.txt
@@ -1,5 +1,0 @@
-target_sources(libcsp PRIVATE
-  csp_rdp.c
-  csp_rdp_queue.c
-  csp_udp.c
-  )
