--- src/transport/csp_udp.c
@@ -1,31 +1,0 @@
-
-
-#include "csp_transport.h"
-
-#include <csp/csp.h>
-#include <csp/csp_debug.h>
-#include <csp/arch/csp_queue.h>
-
-#include "../csp_conn.h"
-
-void csp_udp_new_packet(csp_conn_t * conn, csp_packet_t * packet) {
-
-	/* Enqueue */
-	if (csp_conn_enqueue_packet(conn, packet) < 0) {
-		csp_dbg_conn_ovf++;
-		csp_buffer_free(packet);
-		return;
-	}
-
-	/* Try to queue up the new connection pointer */
-	if (conn->dest_socket != NULL) {
-		if (csp_queue_enqueue(conn->dest_socket->rx_queue, &conn, 0) != CSP_QUEUE_OK) {
-			csp_dbg_conn_ovf++;
-			csp_close(conn);
-			return;
-		}
-
-		/* Ensure that this connection will not be posted to this socket again */
-		conn->dest_socket = NULL;
-	}
-}
